+++ CHANGES.md
@@ -1,5 +1,7 @@
 Change Log
 ==========
+### 2021-10-14
+* Update terriajs and configuration.
 ### 2021-08-30
 * Upgrade to v8 version of terriajs.
 * Update catalogue.